Add vaildErrorHandler
