--- src/main/java/com/moment/the/exception/handler/ValidationErrorHandler.java
@@ -1,33 +1,0 @@
-package com.moment.the.exception.handler;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.validation.BindingResult;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@Slf4j
-@RestControllerAdvice
-public class ValidationErrorHandler {
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public String validationError(MethodArgumentNotValidException exception){
-        BindingResult bindingResult = exception.getBindingResult();
-
-        StringBuilder stringBuilder = new StringBuilder();
-        for (FieldError fieldError : bindingResult.getFieldErrors()){
-            stringBuilder
-                    .append("[")
-                    .append(fieldError.getField())
-                    .append("](은)는 ")
-                    .append(fieldError.getDefaultMessage())
-                    .append(" 입력된 값: [")
-                    .append(fieldError.getRejectedValue())
-                    .append("]");
-        }
-
-        return stringBuilder.toString();
-    }
-
-}
